refactor: package names and casing
we've decided to move to snake casing for package names, inspired by all that
is published on npm.

for components the files are snake cased with the components in pascal case,
this also sings to our unix way of life
